refactor(omegaform): slim useOmegaForm to orchestration
