+++ packages/vue-components/src/components/OmegaForm/types.ts
@@ -12,12 +12,15 @@ import type {
   StandardSchemaV1,
   VueFormApi
 } from "@tanstack/vue-form"
-import type { Effect } from "effect-app"
+import type { Effect, Order, S } from "effect-app"
+import type { UnionToTuples } from "effect-app/utils"
 import type { Fiber as EffectFiber } from "effect/Fiber"
 import type { Redacted } from "effect/Redacted"
-import type { OmegaFieldInternalApi } from "./InputProps"
-import type { MetaRecord } from "./meta/types"
-import type { OF, OmegaFormReturn } from "./useOmegaForm"
+import type { ComputedRef } from "vue"
+import type { MergedInputProps, OmegaFieldInternalApi, TaggedUnionOptionsArray } from "./InputProps"
+import type { MetaRecord, NestedKeyOf } from "./meta/types"
+import type { defaultValuesPriorityUnion as PersistencyPriority, Policies as PersistencyPolicies } from "./persistency"
+import type { FormErrors } from "./submit"
 
 export type FieldPath<T> = unknown extends T ? string
   // technically we cannot have primitive at the root
@@ -246,3 +249,505 @@ export type OmegaAutoGenMeta<
   To extends Record<PropertyKey, any>,
   Name extends DeepKeys<From>
 > = Omit<OmegaInputProps<From, To, Name>, "form">
+
+type keysRule<T> =
+  | {
+    keys?: NestedKeyOf<T>[]
+    banKeys?: "You should only use one of banKeys or keys, not both, moron"
+  }
+  | {
+    keys?: "You should only use one of banKeys or keys, not both, moron"
+    banKeys?: NestedKeyOf<T>[]
+  }
+
+export type Policies = PersistencyPolicies
+export type defaultValuesPriorityUnion = PersistencyPriority
+
+export type OmegaConfig<T> = {
+  i18nNamespace?: string
+
+  persistency?: {
+    /** Order of importance:
+     * - "querystring": Highest priority when persisting
+     * - "local" and then "session": Lower priority storage options
+     */
+    policies?: UnionToTuples<Policies>
+    overrideDefaultValues?: "deprecated: use defaultValuesPriority"
+    id?: string
+  } & keysRule<T>
+
+  ignorePreventCloseEvents?: boolean
+
+  /**
+   * Prevents browser window/tab exit when form has unsaved changes.
+   * Shows native browser "Leave site?" dialog.
+   *
+   * @remarks
+   * - Opt-in only: Must explicitly enable
+   * - Independent from data persistence feature
+   */
+  preventWindowExit?: "prevent" | "prevent-and-reset" | "nope"
+
+  input?: any
+
+  /**
+   * Default values order is: Tanstack default values passed as second parameter to useOmegaForm, then persistency
+   * default values from querystring or local/session storage, then defaults from schema
+   * You can customize the order and  with omegaConfig.defaultValuesPriority
+   * default value = ['tanstack', 'persistency', 'schema']
+   */
+  defaultValuesPriority?: UnionToTuples<defaultValuesPriorityUnion>
+
+  defaultFromSchema?: "deprecated: use defaultValuesPriority"
+}
+
+export interface OF<From, To> extends OmegaFormApi<From, To> {
+  meta: MetaRecord<From>
+  unionMeta: Record<string, MetaRecord<From>>
+  clear: () => void
+  i18nNamespace?: string
+  ignorePreventCloseEvents?: boolean
+  registerField: (
+    field: ComputedRef<{
+      name: string
+      label: string
+      id: string
+    }>
+  ) => void
+  /** @experimental */
+  handleSubmitEffect: {
+    /**
+     * when `checkErrors` is true, the Effect will fail with `FormErrors<From>` when there are validation errors
+     * @experimental */
+    (options: { checkErrors: true; meta?: Record<string, any> }): Effect.Effect<void, FormErrors<From>>
+    /** @experimental */
+    (options?: { meta?: Record<string, any> }): Effect.Effect<void>
+  }
+}
+
+type __VLS_PrettifyLocal<T> =
+  & {
+    [K in keyof T]: T[K]
+  }
+  & {}
+
+// Type aliases for Array component slots - using cached types for performance
+type CachedFieldApi<From, To, TypeProps = DefaultTypeProps> = import("@tanstack/vue-form").FieldApi<
+  From,
+  OmegaFormReturn<From, To, TypeProps>["_keys"],
+  DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>,
+  | import("@tanstack/vue-form").FieldValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldAsyncValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldAsyncValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldAsyncValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldAsyncValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  import("@tanstack/vue-form").FormValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").StandardSchemaV1<From, To>,
+  import("@tanstack/vue-form").FormValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormAsyncValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormAsyncValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormAsyncValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormAsyncValidateOrFn<From> | undefined,
+  Record<string, any> | undefined
+>
+
+type CachedFieldState<From, To, TypeProps = DefaultTypeProps> = import("@tanstack/vue-form").FieldState<
+  From,
+  OmegaFormReturn<From, To, TypeProps>["_keys"],
+  DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>,
+  | import("@tanstack/vue-form").FieldValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldAsyncValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldAsyncValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldAsyncValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  | import("@tanstack/vue-form").FieldAsyncValidateOrFn<
+    From,
+    OmegaFormReturn<From, To, TypeProps>["_keys"],
+    DeepValue<From, OmegaFormReturn<From, To, TypeProps>["_keys"]>
+  >
+  | undefined,
+  import("@tanstack/vue-form").FormValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").StandardSchemaV1<From, To>,
+  import("@tanstack/vue-form").FormValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormAsyncValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormAsyncValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormValidateOrFn<From> | undefined,
+  import("@tanstack/vue-form").FormAsyncValidateOrFn<From> | undefined
+>
+
+export interface OmegaFormReturn<
+  From extends Record<PropertyKey, any>,
+  To extends Record<PropertyKey, any>,
+  TypeProps = DefaultTypeProps
+> extends OF<From, To> {
+  // Pre-computed type aliases - computed ONCE for performance
+  _paths: FieldPath<From>
+  _keys: NestedKeyOf<From>
+  _schema: S.Codec<To, From, never>
+
+  // this crazy thing here is copied from the OmegaFormInput.vue.d.ts, with `From` removed as Generic, instead closed over from the From generic above..
+  Input: <Name extends OmegaFormReturn<From, To, TypeProps>["_paths"]>(
+    __VLS_props: NonNullable<Awaited<typeof __VLS_setup>>["props"],
+    __VLS_ctx?: __VLS_PrettifyLocal<Pick<NonNullable<Awaited<typeof __VLS_setup>>, "attrs" | "emit" | "slots">>,
+    __VLS_expose?: NonNullable<Awaited<typeof __VLS_setup>>["expose"],
+    __VLS_setup?: Promise<{
+      props:
+        & __VLS_PrettifyLocal<
+          & Pick<
+            & Partial<{}>
+            & Omit<
+              {} & import("vue").VNodeProps & import("vue").AllowedComponentProps & import("vue").ComponentCustomProps,
+              never
+            >,
+            never
+          >
+          & TypeProps
+          & Partial<{}>
+        >
+        & BaseProps<From, Name>
+        & import("vue").PublicProps
+      expose(exposed: import("vue").ShallowUnwrapRef<{}>): void
+      attrs: any
+      slots: {
+        default?(props: MergedInputProps<From, Name>): void
+        label?: (props: { required: boolean; id: string; label: string }) => void
+      }
+      emit: {}
+    }>
+  ) => import("vue").VNode & {
+    __ctx?: Awaited<typeof __VLS_setup>
+  }
+  Errors: (
+    __VLS_props: NonNullable<Awaited<typeof __VLS_setup>>["props"],
+    __VLS_ctx?: __VLS_PrettifyLocal<Pick<NonNullable<Awaited<typeof __VLS_setup>>, "attrs" | "emit" | "slots">>,
+    __VLS_expose?: NonNullable<Awaited<typeof __VLS_setup>>["expose"],
+    __VLS_setup?: Promise<{
+      props:
+        & __VLS_PrettifyLocal<
+          & Pick<
+            & Partial<{}>
+            & Omit<
+              {} & import("vue").VNodeProps & import("vue").AllowedComponentProps & import("vue").ComponentCustomProps,
+              never
+            >,
+            never
+          >
+          & Partial<{}>
+        >
+        & import("vue").PublicProps
+      expose(exposed: import("vue").ShallowUnwrapRef<{}>): void
+      attrs: any
+      slots: {
+        default: (props: { errors: readonly OmegaError[]; showedGeneralErrors: string[] }) => void
+      }
+      emit: {}
+    }>
+  ) => import("vue").VNode & {
+    __ctx?: Awaited<typeof __VLS_setup>
+  }
+  TaggedUnion: <Name extends OmegaFormReturn<From, To, TypeProps>["_keys"]>(
+    __VLS_props: NonNullable<Awaited<typeof __VLS_setup>>["props"],
+    __VLS_ctx?: __VLS_PrettifyLocal<Pick<NonNullable<Awaited<typeof __VLS_setup>>, "attrs" | "emit" | "slots">>,
+    __VLS_expose?: NonNullable<Awaited<typeof __VLS_setup>>["expose"],
+    __VLS_setup?: Promise<{
+      props:
+        & __VLS_PrettifyLocal<
+          & Pick<
+            & Partial<{}>
+            & Omit<
+              {} & import("vue").VNodeProps & import("vue").AllowedComponentProps & import("vue").ComponentCustomProps,
+              never
+            >,
+            never
+          >
+          & {
+            name?: Name
+            type?: "select" | "radio"
+            options: TaggedUnionOptionsArray<From, Name>
+            _debugName?: [NoInfer<Name>]
+            label?: string
+          }
+          & {}
+        >
+        & import("vue").PublicProps
+      expose(exposed: import("vue").ShallowUnwrapRef<{}>): void
+      attrs: any
+      slots: Record<
+        string,
+        (props: {
+          field: import("@tanstack/vue-form").FieldApi<
+            From,
+            Name,
+            DeepValue<From, Name>,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any
+          >
+          state: import("@tanstack/vue-form").FieldState<
+            From,
+            Name,
+            DeepValue<From, Name>,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any,
+            any
+          >
+        }) => any
+      >
+      emit: {}
+    }>
+  ) => import("vue").VNode & {
+    __ctx?: Awaited<typeof __VLS_setup>
+  }
+  Array: <Name extends OmegaFormReturn<From, To, TypeProps>["_keys"]>(
+    __VLS_props: NonNullable<Awaited<typeof __VLS_setup>>["props"],
+    __VLS_ctx?: __VLS_PrettifyLocal<Pick<NonNullable<Awaited<typeof __VLS_setup>>, "attrs" | "emit" | "slots">>,
+    __VLS_expose?: NonNullable<Awaited<typeof __VLS_setup>>["expose"],
+    __VLS_setup?: Promise<{
+      props:
+        & __VLS_PrettifyLocal<
+          & Pick<
+            & Partial<{}>
+            & Omit<
+              {} & import("vue").VNodeProps & import("vue").AllowedComponentProps & import("vue").ComponentCustomProps,
+              never
+            >,
+            never
+          >
+          & (Omit<OmegaArrayProps<From, To, Name>, "form">)
+          & {}
+        >
+        & import("vue").PublicProps
+      expose(exposed: import("vue").ShallowUnwrapRef<{}>): void
+      attrs: any
+      slots: {
+        "pre-array"?: (props: {
+          field: CachedFieldApi<From, To, TypeProps>
+          state: CachedFieldState<From, To, TypeProps>
+        }) => any
+      } & {
+        default?: (props: {
+          subField: CachedFieldApi<From, To, TypeProps>
+          subState: CachedFieldState<From, To, TypeProps>
+          index: number
+          field: CachedFieldApi<From, To, TypeProps>
+        }) => any
+      } & {
+        "post-array"?: (props: {
+          field: CachedFieldApi<From, To, TypeProps>
+          state: CachedFieldState<From, To, TypeProps>
+        }) => any
+      } & {
+        field?: (props: {
+          field: CachedFieldApi<From, To, TypeProps>
+        }) => any
+      }
+      emit: {}
+    }>
+  ) => import("vue").VNode & {
+    __ctx?: Awaited<typeof __VLS_setup>
+  }
+
+  AutoGen: <Name extends OmegaFormReturn<From, To, TypeProps>["_keys"]>(
+    __VLS_props: NonNullable<Awaited<typeof __VLS_setup>>["props"],
+    __VLS_ctx?: __VLS_PrettifyLocal<Pick<NonNullable<Awaited<typeof __VLS_setup>>, "attrs" | "emit" | "slots">>,
+    __VLS_expose?: NonNullable<Awaited<typeof __VLS_setup>>["expose"],
+    __VLS_setup?: Promise<{
+      props:
+        & __VLS_PrettifyLocal<
+          Pick<
+            & Partial<{}>
+            & Omit<
+              {} & import("vue").VNodeProps & import("vue").AllowedComponentProps & import("vue").ComponentCustomProps,
+              never
+            >,
+            never
+          > & {
+            // form: OmegaInputProps<From, To>["form"]
+            pick?: OmegaFormReturn<From, To, TypeProps>["_keys"][]
+            omit?: OmegaFormReturn<From, To, TypeProps>["_keys"][]
+            labelMap?: (key: OmegaFormReturn<From, To, TypeProps>["_keys"]) => string | undefined
+            filterMap?: <M extends OmegaAutoGenMeta<From, To, Name>>(
+              key: OmegaFormReturn<From, To, TypeProps>["_keys"],
+              meta: M
+            ) => boolean | M
+            order?: OmegaFormReturn<From, To, TypeProps>["_keys"][]
+            sort?: Order.Order<OmegaAutoGenMeta<From, To, Name>>
+          } & {}
+        >
+        & import("vue").PublicProps
+      expose(exposed: import("vue").ShallowUnwrapRef<{}>): void
+      attrs: any
+      slots: {
+        default(props: {
+          child: OmegaAutoGenMeta<From, To, Name>
+        }): void
+      }
+      emit: {}
+    }>
+  ) => import("vue").VNode & {
+    __ctx?: Awaited<typeof __VLS_setup>
+  }
+
+  Form: <K extends keyof OmegaFormState<To, From>>(
+    __VLS_props: NonNullable<Awaited<typeof __VLS_setup>>["props"],
+    __VLS_ctx?: __VLS_PrettifyLocal<Pick<NonNullable<Awaited<typeof __VLS_setup>>, "attrs" | "emit" | "slots">>,
+    __VLS_expose?: NonNullable<Awaited<typeof __VLS_setup>>["expose"],
+    __VLS_setup?: Promise<{
+      props:
+        & __VLS_PrettifyLocal<
+          Pick<
+            & Partial<{}>
+            & Omit<
+              {} & import("vue").VNodeProps & import("vue").AllowedComponentProps & import("vue").ComponentCustomProps,
+              never
+            >,
+            never
+          > & {
+            // form: OmegaFormReturn<From, To, Props>
+            disabled?: boolean
+            subscribe?: K[]
+          } & {}
+        >
+        & import("vue").PublicProps
+      expose(exposed: import("vue").ShallowUnwrapRef<{}>): void
+      attrs: any
+      slots: {
+        default(props: {
+          subscribedValues: K[] extends undefined[] ? Record<string, never> : Pick<OmegaFormState<From, To>, K>
+        }): void
+      }
+      emit: {}
+    }>
+  ) => import("vue").VNode & {
+    __ctx?: Awaited<typeof __VLS_setup>
+  }
+}
